% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.stat.diff.test.R
\name{calculate.p.value}
\alias{calculate.p.value}
\title{Calculate p-value}
\usage{
calculate.p.value(data, paired = FALSE, type = c("parametric",
  "non-parametric"))
}
\arguments{
\item{data}{data frame}

\item{paired}{a logical indicating whether you want a paired test.}

\item{type}{select test type whether parametric or non-parametric.}
}
\description{
Calculate p-value
}
