% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.stat.diff.test.R
\name{ez.check.density}
\alias{ez.check.density}
\title{Check probability density}
\usage{
ez.check.density(data, group.name = "group", value.name = "value")
}
\arguments{
\item{data}{response vector.}

\item{group.name}{a name for the group (default: "group")}

\item{value.name}{a name for the value (default: "value")}
}
\description{
Compute and plot probability density.
}
\examples{
{
A <- c(1, 2, 1, 1, 1, 1, 1, 1, 1, 1, 2)
B <- c(3, 3, 4, 3, 1, 2, 3, 1, 1, 5, 4)
data <- data.frame(A,B)

ez.check.density(data)
}

}
