% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.stat.diff.test.R
\name{plot.graph}
\alias{plot.graph}
\title{Plot graph}
\usage{
\method{plot}{graph}(data, type = c("parametric", "non-parametric"))
}
\arguments{
\item{data}{data frame}

\item{type}{parametric of non-parametric}
}
\description{
Plot graph
}
