% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.stat.diff.test.R
\name{adjust.p.value}
\alias{adjust.p.value}
\title{Adjust p-value for multiple comparisons}
\usage{
adjust.p.value(p.value, adjust.method = c("holm", "bonferroni", "BH"))
}
\arguments{
\item{p.value}{numeric vector of p-values.}

\item{adjust.method}{method for adjusting p-values.}
}
\description{
Given a set of p-values, returns adjusted p-values.
}
