% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.stat.diff.test.R
\name{adjust.p.value}
\alias{adjust.p.value}
\title{Adjust p-value}
\usage{
adjust.p.value(p.value, type = c("parametric", "non-parametric"),
  adjust.method = c("holm", "bonferroni", "BH"))
}
\arguments{
\item{p.value}{p-value to adjust}

\item{type}{parametric of non-parametric}

\item{adjust.method}{p-values adjusted using several methods}
}
\description{
Adjust p-value
}
