% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.stat.diff.test.R
\name{ez.stat.diff.test}
\alias{ez.stat.diff.test}
\title{Conduct statistical difference tests}
\usage{
ez.stat.diff.test(data, type = c("parametric", "non-parametric"),
  paired = FALSE, adjust.method = c("holm", "bonferroni", "BH"),
  plot = TRUE, group.name = "group", value.name = "value")
}
\arguments{
\item{data}{response vector.}

\item{type}{select test type whether parametric or non-parametric.}

\item{paired}{a logical indicating whether you want a paired test.}

\item{adjust.method}{method for adjusting p-values.}

\item{plot}{a logical indicating whether you want a plot graph.}

\item{group.name}{a name for the group (default: "group")}

\item{value.name}{a name for the value (default: "value")}
}
\description{
Calculate two sample or pairwise comparisons with corrections.
}
\examples{
{
A <- c(1, 2, 1, 1, 1, 1, 1, 1, 1, 1, 2)
B <- c(3, 3, 4, 3, 1, 2, 3, 1, 1, 5, 4)
data <- data.frame(A,B)

ez.stat.diff.test(data, "parametric", paired=TRUE, adjust.method = "holm")
}
}
