% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.stat.diff.test.R
\name{ez.stat.diff.test}
\alias{ez.stat.diff.test}
\title{Conduct statistical difference test}
\usage{
ez.stat.diff.test(data, type = c("parametric", "non-parametric"),
  paired = FALSE, adjust.method = c("holm", "bonferroni", "BH"),
  plot = TRUE)
}
\arguments{
\item{data}{data frame}

\item{type}{parametric of non-parametric.}

\item{paired}{a logical indicating whether you want a paired test.}

\item{adjust.method}{p-values adjusted using several methods.}

\item{plota}{a logical indicating whether you want a plot graph.}
}
\description{
Conduct statistical difference test
}
