% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.stat.diff.test.R
\name{plot.graph}
\alias{plot.graph}
\title{Plot graph}
\usage{
\method{plot}{graph}(data, type = c("parametric", "non-parametric"),
  group.name = "group", value.name = "value")
}
\arguments{
\item{data}{response vector.}

\item{type}{select test type whether parametric or non-parametric.}

\item{group.name}{a name for the group (default: "group")}

\item{value.name}{a name for the value (default: "value")}
}
\description{
Plot bar graphs (parametric case) or box plots (non-parametric case).
}
