% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.stat.diff.test.R
\name{plot.graph}
\alias{plot.graph}
\title{Plot graph}
\usage{
\method{plot}{graph}(data, type = c("parametric", "non-parametric"))
}
\arguments{
\item{data}{response vector.}

\item{type}{select test type whether parametric or non-parametric.}
}
\description{
Plot bar graphs (parametric case) or box plots (non-parametric case).
}
